octra-app/requirements.txt
aiohttp
pynacl

